requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Cython==0.29
docutils==0.14
Flask==1.0.2
FoLiA-tools==1.5.1.60
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
pkg-resources==0.0.0
PyNLPl==1.2.7
pyparsing==2.2.2
python-ucto==0.4.7
PyYAML==3.13
rdflib==4.2.2
requests==2.19.1
six==1.11.0
urllib3==1.23
webencodings==0.5.1
Werkzeug==0.14.1
